Add option to isplay horizontal distance in hours
The Horus team found that horizontal distances in m/s were hard to get
an intuition for, particularly as the speed across land is often used
for calculating if a car can travel to the landing site in time.

This change adds a selectable option that modifies the horizontal speed
to be calculated in km/h, or m/h if the user has selected imperial
units. The default behaviour is still to use distance per second.

Signed-off-by: Joel Stanley <joel@jms.id.au>